Reduce object and iteration number.
Do some manual log inspection before proceed on log processing.

+++ mem-dummy.c
@@ -10,4 +10,4 @@ void mem_write(int32_t *addr, int32_t val) {
 }
 
 void mem_init(int nthr) {}
-void mem_init_thr(long tid) {}
+void mem_init_thr(int tid) {}
